// src/widgets/report/types.ts
export enum DragType {
	NONE = 'none',
	DND = 'drag-drop',
	RESIZE_TOP = 'resize-top',
	RESIZE_LEFT = 'resize-left',
	RESIZE_RIGHT = 'resize-right',
	RESIZE_BOTTOM = 'resize-bottom',
	RESIZE_TOP_LEFT = 'resize-top-left',
	RESIZE_TOP_RIGHT = 'resize-top-right',
	RESIZE_BOTTOM_LEFT = 'resize-bottom-left',
	RESIZE_BOTTOM_RIGHT = 'resize-bottom-right'
}